X86_64-linux Dockerfile troubleshooting
Currently I am building this project on an M1. However, Docker has a
hard time building from the M1 Max CPU architecture to a x86_64-linux
target (the target that the ruby:3.0.4-alpine image uses, I believe),
and dependecies are not getting installed correctly. Specifically
either nokogiri (which unfortunately I created this rails app with,
and in hindsight would have excluded actiontext to prevent this) and
sqlite (no idea why this isn't working).